move threshold arg parsing into cli/utils
